fix: keep scheduled tasks from drawing unrelated skill guidance (#6608)
