+++ server/lib/index.js
@@ -384,6 +384,7 @@ export * from './taskBlockCategories.js';
 export * from './taskRequeue.js';
 export * from './taskRetryHold.js';
 export * from './taskTargetBranch.js';
+export * from './scheduledTaskTypes.js';
 export * from './taskTargetScope.js';
 export * from './taxonomyTally.js';
 export * from './worktreeOwnership.js';
